remove: 불필요 객체 및 테스트 삭제
